% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acs14.R
\name{acs14}
\alias{acs14}
\title{Fetch data from the 2010-2014 American Community Survey}
\usage{
acs14(api_key = NULL, geography = "us", variable = "B01001_001E",
  state = NULL, county = NULL)
}
\arguments{
\item{api_key}{The Census API key of the user; can be set beforehand with the set_api_key function.}

\item{geography}{The geography for which the user would like to request data.  Available geographies include 'us', 'region', 'division', 'state', 'county', 'tract', and 'block group'.  Defaults to 'us'.}

\item{variable}{A character string or vector of character strings representing the variable name(s) for which the user would like to request data.  Defaults to 'B01001_001E', the estimate for Total Population.}

\item{state}{The state for which data will be requested.  Defaults to NULL.}

\item{county}{The county for which data will be requested.  Defaults to NULL.}
}
\description{
Lightweight interface to the US Census Bureau's API for the 2010-2014 American Community Survey.
}

