% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/acs14_package.R
\docType{package}
\name{acs14lite}
\alias{acs14lite}
\alias{acs14lite-package}
\title{Access 2010-2014 American Community Survey data from R}
\description{
Access 2010-2014 American Community Survey data from R
}
\author{
Kyle Walker
}

