% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/moe.R
\name{moe_ratio}
\alias{moe_ratio}
\title{Calculate the margin of error for a derived ratio}
\usage{
moe_ratio(num, denom, moe_num, moe_denom)
}
\description{
Calculate the margin of error for a derived ratio
}

